Update Rakefile to check for linked files in task :uninstall

+++ Rakefile
@@ -131,8 +131,10 @@ end
 
 task :uninstall do
   locals_listing.each do |f|
-    puts "Warn!  | Unlink    | file: #{f[:new_path]}"
-    File.unlink f[:new_path]
+    if current_file(f)[:type] == :symlink
+      puts "Warn!  | Unlink    | file: #{f[:new_path]}"
+      File.unlink f[:new_path]
+    end
   end
   dotfile_listing.each do |f|
     puts "Warn!  | Unlink    | file: #{f[:new_path]}"
